requirements.txt
python-pptx
openpyxl
tqdm
PyYAML